added cool app icon

+++ composeApp/src/commonMain/kotlin/uz/coder/foottopbusiness/data/network/dto/admin/DailyRevenue.kt
@@ -9,5 +9,5 @@ data class DailyRevenue(
     @SerialName("date")
     val date: String? = null,
     @SerialName("revenue")
-    val revenue: Int? = null
+    val revenue: Double? = null
 )
