+++ composeApp/src/commonMain/kotlin/uz/coder/foottopbusiness/domain/model/DailyRevenue.kt
@@ -2,5 +2,5 @@ package uz.coder.foottopbusiness.domain.model
 
 data class DailyRevenue(
     val date: String,
-    val revenue: Int
+    val revenue: Double
 )
